# backend/src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Clientes",
    "description": "Documentação da API CRUD para as tabelas de clientes com Node.js e Supabase.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor local"
    }
  ],
  "tags": [
    {
      "name": "Cliente",
      "description": "Endpoints para gerenciar clientes"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/cliente/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/cliente/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Cliente": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "id": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  },
                  "description": {
                    "type": "string",
                    "example": "ID do cliente"
                  },
                  "example": {
                    "type": "number",
                    "example": 1
                  }
                }
              },
              "created_at": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "date-time"
                  },
                  "description": {
                    "type": "string",
                    "example": "Data de criação"
                  },
                  "example": {
                    "type": "string",
                    "example": "2025-09-17T12:00:00Z"
                  }
                }
              },
              "nome": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "description": {
                    "type": "string",
                    "example": "Nome do cliente"
                  },
                  "example": {
                    "type": "string",
                    "example": "Empresa XYZ"
                  }
                }
              },
              "cnpj": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "description": {
                    "type": "string",
                    "example": "CNPJ do cliente"
                  },
                  "example": {
                    "type": "string",
                    "example": "12345678000195"
                  }
                }
              },
              "segmento": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "description": {
                    "type": "string",
                    "example": "Segmento do cliente"
                  },
                  "example": {
                    "type": "string",
                    "example": "Varejo"
                  }
                }
              },
              "cep": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "description": {
                    "type": "string",
                    "example": "CEP do cliente"
                  },
                  "example": {
                    "type": "string",
                    "example": "12345678"
                  }
                }
              },
              "endereco": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "description": {
                    "type": "string",
                    "example": "Endereço do cliente"
                  },
                  "example": {
                    "type": "string",
                    "example": "Rua Principal"
                  }
                }
              },
              "numero": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "description": {
                    "type": "string",
                    "example": "Número do endereço"
                  },
                  "example": {
                    "type": "string",
                    "example": "100"
                  }
                }
              },
              "bairro": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "description": {
                    "type": "string",
                    "example": "Bairro do cliente"
                  },
                  "example": {
                    "type": "string",
                    "example": "Centro"
                  }
                }
              },
              "cidade": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "description": {
                    "type": "string",
                    "example": "Cidade do cliente"
                  },
                  "example": {
                    "type": "string",
                    "example": "São Paulo"
                  }
                }
              },
              "estado": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "description": {
                    "type": "string",
                    "example": "Estado do cliente (UF)"
                  },
                  "example": {
                    "type": "string",
                    "example": "SP"
                  }
                }
              }
            }
          }
        }
      },
      "ClienteInput": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "required": {
            "type": "array",
            "example": [
              "nome",
              "cnpj",
              "segmento",
              "cep",
              "endereco",
              "numero",
              "bairro",
              "cidade",
              "estado"
            ],
            "items": {
              "type": "string"
            }
          },
          "properties": {
            "type": "object",
            "properties": {
              "nome": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "description": {
                    "type": "string",
                    "example": "Nome do cliente"
                  },
                  "example": {
                    "type": "string",
                    "example": "Empresa XYZ"
                  }
                }
              },
              "cnpj": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "description": {
                    "type": "string",
                    "example": "CNPJ do cliente"
                  },
                  "example": {
                    "type": "string",
                    "example": "12345678000195"
                  }
                }
              },
              "segmento": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "description": {
                    "type": "string",
                    "example": "Segmento do cliente"
                  },
                  "example": {
                    "type": "string",
                    "example": "Varejo"
                  }
                }
              },
              "cep": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "description": {
                    "type": "string",
                    "example": "CEP do cliente"
                  },
                  "example": {
                    "type": "string",
                    "example": "12345678"
                  }
                }
              },
              "endereco": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "description": {
                    "type": "string",
                    "example": "Endereço do cliente"
                  },
                  "example": {
                    "type": "string",
                    "example": "Rua Principal"
                  }
                }
              },
              "numero": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "description": {
                    "type": "string",
                    "example": "Número do endereço"
                  },
                  "example": {
                    "type": "string",
                    "example": "100"
                  }
                }
              },
              "bairro": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "description": {
                    "type": "string",
                    "example": "Bairro do cliente"
                  },
                  "example": {
                    "type": "string",
                    "example": "Centro"
                  }
                }
              },
              "cidade": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "description": {
                    "type": "string",
                    "example": "Cidade do cliente"
                  },
                  "example": {
                    "type": "string",
                    "example": "São Paulo"
                  }
                }
              },
              "estado": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "description": {
                    "type": "string",
                    "example": "Estado do cliente (UF)"
                  },
                  "example": {
                    "type": "string",
                    "example": "SP"
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}